Reorganize pet hierarchy in expansion order
Fix #140

The previous hierarchy was really messy. The top-level categories are
often useful only for one expansion and then it's hard to fit new pet
obtention "paradigms" without creating new categories.

The per-expansion approach is a lot more flexible, and is what is used
for all the other categories (mounts, toys, wild pets, ...). So this
hierarchy is a lot more consistent with the rest of the website.

This also allowed to split/move a lot of categories that didn't make
sense, and to fix a lot of pets that were in completely random places.

This probably isn't perfect and a lot of pets are still missing, but
it's a vast improvement that will allow us to better find and add the
missing pets.