toys: rename side/allowableClasses attributes
Since the toys are always shown in the toybox whatever the side/class
is, the only way for the exact count to be correct is to remove all
those attributes.

Since one day we might start to want filters to hide toys that cannot be
obtained on a specific character, I just renamed the attributes instead
of removing them.